updated tests with struct member datatypes (#8)
used updated code generator to generate tests that have struct member datatypes specified when possible

+++ test/petstore_v2/petstore_test_storeapi.jl
@@ -12,7 +12,7 @@ function test(uri)
 
     @info("StoreApi - getInventory")
     inventory = getInventory(api)
-    @test isa(inventory, Dict{String,Int32})
+    @test isa(inventory, Dict{String,Int64})
     @test !isempty(inventory)
 
     @info("StoreApi - placeOrder")
